feat: fix single result with when there is quiz been deleted situation.

+++ migrations/20231221061845-create-quiz.js
@@ -25,7 +25,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       answer: {
-        type: Sequelize.TINYINT
+        type: Sequelize.SMALLINT
       },
       is_delete: {
         type: Sequelize.BOOLEAN
